#58 major changes in many classes
What: fixed checkstyle in all relevant classes. Added new method
to storage to update price of clothing.
Also fixed restserver-controller a bit, and added new methods
to both client and restserver.

Why: everything was needed to test out the rest API. This is an experiment
branch, so I had to do it to see how everything works